requirements.txt
torch==1.1.0
numpy==1.17.0
torchvision==0.3.0
Pillow==6.1.0
